backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==0.8.11
PyMuPDF==1.23.26
textstat==0.7.3
numpy>=1.26.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# AI Enhancement Dependencies
huggingface-hub==0.20.3
httpx==0.26.0
tenacity==8.2.3
cachetools==5.3.2

# Job Verification Dependencies
python-whois==0.8.0
pytesseract==0.3.10
pillow==10.1.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
